summary/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
azure-ai-textanalytics==5.3.0b2